backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
google-generativeai
google-cloud-documentai
python-docx
PyPDF2